backend/requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional bonus features (install as needed)
transformers>=4.35.0
torch>=2.1.0
# googletrans is omitted to avoid httpx conflicts; install manually if needed:
# googletrans==4.0.0-rc1

# HTTP requests for HuggingFace fallback
requests>=2.31.0

